requirements.txt
pillow
pillow_heif
click
natosort